Code: Moved layer panel to proper folder structure
