+++ packages/story-editor/src/components/panels/design/index.js
@@ -20,7 +20,6 @@ export { default as CaptionsPanel } from './captions';
 export { default as ElementAlignmentPanel } from './alignment';
 export { default as FilterPanel } from './filter';
 export { default as ImageAccessibilityPanel } from './imageAccessibility';
-export { default as LayerPanel } from './layer';
 export { default as LinkPanel } from './link';
 export { default as PageAttachmentPanel } from './pageAttachment';
 export { default as PageBackgroundPanel } from './pageBackground';
