+++ packages/story-editor/src/components/panels/layer/layer.js
@@ -29,19 +29,26 @@ import {
   THEME_CONSTANTS,
   Input,
 } from '@googleforcreators/design-system';
-import { useRef, memo, useState } from '@googleforcreators/react';
+import { useRef, memo, useState, useEffect } from '@googleforcreators/react';
+import {
+  getDefinitionForType,
+  getLayerName,
+} from '@googleforcreators/elements';
 import { useFeature } from 'flagged';
 
 /**
  * Internal dependencies
  */
-import PropTypes from 'prop-types';
-import { v4 as uuidv4 } from 'uuid';
-import { useCanvas, useStory } from '../../../../app';
-import Tooltip from '../../../tooltip';
-import generateGroupName from '../../../../utils/generateGroupName';
+import StoryPropTypes from '../../../types';
+import { useStory, useCanvas } from '../../../app';
+import useCORSProxy from '../../../utils/useCORSProxy';
+import usePerformanceTracking from '../../../utils/usePerformanceTracking';
+import { TRACKING_EVENTS } from '../../../constants';
+import Tooltip from '../../tooltip';
+import useShapeMask from '../../../utils/useShapeMask';
+import useLayerSelection from './useLayerSelection';
+import ShapeMaskWrapper from './shapeMaskWrapper';
 import { LAYER_HEIGHT, NESTED_PX } from './constants';
-import useGroupSelection from './useGroupSelection';
 
 const fadeOutCss = css`
   background-color: var(--background-color);
@@ -112,7 +119,7 @@ const LayerButton = styled(Button).attrs({
   align-items: center;
   user-select: none;
   border-radius: 0;
-  padding-left: ${({ isNested }) => (isNested ? 30 : 12)}px;
+  padding-left: ${({ isNested }) => (isNested ? NESTED_PX : 12)}px;
   transition: revert;
 
   ${({ isSelected, theme }) =>
@@ -153,51 +160,65 @@ const LayerButton = styled(Button).attrs({
   }
 `;
 
-const GroupIconsWrapper = styled.div`
+const LayerInputWrapper = styled.div`
+  display: grid;
+  grid-template-columns: 23px 1fr;
+  height: 100%;
+  width: 100%;
+  padding-left: ${({ isNested }) => (isNested ? NESTED_PX : 12)}px;
+  padding-right: 10px;
+
+  :hover {
+    background: ${({ theme }) => theme.colors.interactiveBg.tertiaryHover};
+  }
+  :hover,
+  :hover + * {
+    --background-color: ${({ theme }) =>
+      theme.colors.interactiveBg.tertiaryHover};
+    --background-color-opaque: ${({ theme }) =>
+      rgba(theme.colors.interactiveBg.tertiaryHover, 0)};
+  }
+`;
+
+const LayerInput = styled(Input)`
+  overflow: visible;
+
+  div {
+    height: 100%;
+  }
+`;
+
+const LayerIconWrapper = styled.div`
   display: flex;
   align-items: center;
   justify-content: flex-start;
-  width: 48px;
+  color: ${({ theme }) => theme.colors.fg.primary};
 
-  svg {
-    position: relative;
-    display: block;
-    width: 100%;
-    color: ${({ theme }) => theme.colors.fg.secondary};
+  :hover {
+    background: ${({ theme }) => theme.colors.interactiveBg.tertiaryHover};
   }
-
-  /*
-  * moves the click target + positioning to the left
-  * this ensures a better click target for 
-  * the arrow expand collapse state
-  */
-  transform: translateX(-12px);
-`;
-
-const ChevronDown = styled(Icons.ChevronDown)`
-  min-width: 32px;
-  min-height: 32px;
-`;
-
-const ChevronRight = styled(Icons.ChevronDown)`
-  /*
-  * Using ChevronDown and rotating it here 
-  * vs ChevronRightSmall to keep consistent sizing between states
-  */
-  min-width: 32px;
-  min-height: 32px;
-  transform: rotate(-90deg);
+  :hover,
+  :hover + * {
+    --background-color: ${({ theme }) =>
+      theme.colors.interactiveBg.tertiaryHover};
+    --background-color-opaque: ${({ theme }) =>
+      rgba(theme.colors.interactiveBg.tertiaryHover, 0)};
+  }
 `;
 
 const LayerDescription = styled.div`
   position: relative;
-  max-width: 100%;
+  width: 100%;
   overflow: hidden;
   display: flex;
   align-items: center;
-  margin-left: 12px;
+  margin-left: 0;
   text-align: left;
   color: ${({ theme }) => theme.colors.fg.primary};
+`;
+
+const LayerInputForm = styled(LayerDescription).attrs({ as: 'form' })`
+  overflow: visible;
 `;
 
 const LayerText = styled(Text).attrs({
@@ -298,85 +319,71 @@ const LayerAction = styled(Button).attrs({
   }
 `;
 
-const LayerInputWrapper = styled.div`
-  display: grid;
-  grid-template-columns: 23px 1fr;
-  height: 100%;
-  width: 100%;
-  padding-left: ${({ isNested }) => (isNested ? NESTED_PX : 12)}px;
-  padding-right: 10px;
-
-  :hover {
-    background: ${({ theme }) => theme.colors.interactiveBg.tertiaryHover};
-  }
-  :hover,
-  :hover + * {
-    --background-color: ${({ theme }) =>
-      theme.colors.interactiveBg.tertiaryHover};
-    --background-color-opaque: ${({ theme }) =>
-      rgba(theme.colors.interactiveBg.tertiaryHover, 0)};
-  }
-`;
-
-const LayerInputForm = styled(LayerDescription).attrs({ as: 'form' })`
-  overflow: visible;
-`;
-
-const LayerInput = styled(Input)`
-  overflow: visible;
-
-  div {
-    height: 100%;
-  }
-`;
-
 function preventReorder(e) {
   e.stopPropagation();
   e.preventDefault();
 }
 
-function Group({ groupId }) {
+function Layer({ element }) {
+  const layerName = getLayerName(element);
+  const [newLayerName, setNewLayerName] = useState(layerName);
   const isLayerLockingEnabled = useFeature('layerLocking');
+  const { LayerIcon, hasDuplicateMenu } = getDefinitionForType(element.type);
+  const { isSelected, handleClick } = useLayerSelection(element);
+  const { isDefaultBackground } = element;
   const {
+    duplicateElementsById,
+    updateElementById,
+    deleteElementById,
+    currentPageBackgroundColor,
     groups,
-    updateGroupById,
-    deleteGroupAndElementsById,
-    updateElementsById,
-    duplicateGroupById,
-    groupLayers,
   } = useStory(({ actions, state }) => ({
-    groups: state.currentPage.groups,
-    updateGroupById: actions.updateGroupById,
-    deleteGroupAndElementsById: actions.deleteGroupAndElementsById,
-    updateElementsById: actions.updateElementsById,
-    duplicateGroupById: actions.duplicateGroupById,
-    groupLayers: state.currentPage.elements.filter(
-      (el) => el.groupId === groupId
-    ),
+    duplicateElementsById: actions.duplicateElementsById,
+    deleteElementById: actions.deleteElementById,
+    updateElementById: actions.updateElementById,
+    groups: state.currentPage?.groups || {},
+    currentPageBackgroundColor:
+      !isDefaultBackground || state.currentPage?.backgroundColor,
   }));
 
-  const group = groups[groupId];
-  const groupRef = useRef(null);
-  const deleteButtonRef = useRef(null);
-  const groupDomId = `group-${groupId}`;
-
-  const lockTitle = group.isLocked
-    ? __('Unlock Group', 'web-stories')
-    : __('Lock Group', 'web-stories');
-
-  const LockIcon = group.isLocked ? Icons.LockClosed : Icons.LockOpen;
-
-  const layerName = group.name;
-  const [newLayerName, setNewLayerName] = useState(layerName);
-  const { isSelected, handleClick } = useGroupSelection(groupId);
-  const isLayerNamingEnabled = useFeature('layerNaming');
   const { renamableLayer, setRenamableLayer } = useCanvas(
     ({ state, actions }) => ({
       renamableLayer: state.renamableLayer,
       setRenamableLayer: actions.setRenamableLayer,
     })
   );
-  const isRenameable = renamableLayer?.elementId === groupId;
+
+  useEffect(() => {
+    setNewLayerName(layerName);
+  }, [layerName]);
+
+  const { hasShapeMask, removeShapeMask } = useShapeMask(element);
+  const removeMaskTitle = __('Unmask', 'web-stories');
+  const { getProxiedUrl } = useCORSProxy();
+  const layerRef = useRef(null);
+  usePerformanceTracking({
+    node: layerRef.current,
+    eventData: { ...TRACKING_EVENTS.SELECT_ELEMENT, label: element.type },
+  });
+
+  const deleteButtonRef = useRef(null);
+  usePerformanceTracking({
+    node: deleteButtonRef.current,
+    eventData: { ...TRACKING_EVENTS.DELETE_ELEMENT, label: element.type },
+  });
+
+  const layerId = `layer-${element.id}`;
+  const isNested = element.groupId;
+
+  const lockTitle = isNested
+    ? element.isLocked
+      ? __('Group is Locked', 'web-stories')
+      : __('Group is Unlocked', 'web-stories')
+    : element.isLocked
+    ? __('Unlock Layer', 'web-stories')
+    : __('Lock Layer', 'web-stories');
+
+  const LockIcon = element.isLocked ? Icons.LockClosed : Icons.LockOpen;
 
   const handleChange = (evt) => {
     setNewLayerName(evt.target.value);
@@ -392,17 +399,11 @@ function Group({ groupId }) {
   const updateLayerName = () => {
     setRenamableLayer(null);
     const trimmedLayerName = newLayerName.trim();
-    // Don't update name if trimmed layer name is empty.
-    // This means that submitting an empty name will exit renaming, and the
-    // layer name will revert to whatever it was before, ignoring the empty input.
-    if (!trimmedLayerName) {
-      setNewLayerName(layerName);
-    } else {
-      updateGroupById({
-        groupId,
-        properties: {
-          name: trimmedLayerName,
-        },
+    // Only update name if trimmed layer name is not empty.
+    if (trimmedLayerName) {
+      updateElementById({
+        elementId: element.id,
+        properties: { layerName: trimmedLayerName },
       });
     }
   };
@@ -417,49 +418,24 @@ function Group({ groupId }) {
   // move focus, which will blur the input, which will cancel renaming.
   const stopPropagation = (evt) => evt.stopPropagation();
 
-  const handleLockGroup = () => {
-    updateGroupById({
-      groupId,
-      properties: { isLocked: !group.isLocked },
-    });
-    updateElementsById({
-      elementIds: groupLayers.map((layer) => layer.id),
-      properties: { isLocked: !group.isLocked },
-    });
-  };
+  const isLayerNamingEnabled = useFeature('layerNaming');
+  const isRenameable = renamableLayer?.elementId === element.id;
+  const group = groups[element.groupId];
 
-  const handleDuplicateGroup = () => {
-    const newGroupId = uuidv4();
-    const newGroupName = generateGroupName(groups, group.name);
-    duplicateGroupById({
-      oldGroupId: groupId,
-      groupId: newGroupId,
-      name: newGroupName,
-      isLocked: group.isLocked,
-      isCollapsed: group.isCollapsed,
-    });
-  };
-  const handleGroupArrowClick = () => {
-    updateGroupById({
-      groupId,
-      properties: { isCollapsed: !group.isCollapsed },
-    });
-  };
-  const isCollapsed = group.isCollapsed;
-  const GroupArrow = isCollapsed ? (
-    <ChevronRight onClick={handleGroupArrowClick} />
-  ) : (
-    <ChevronDown onClick={handleGroupArrowClick} />
-  );
-
-  return (
+  // Done like this because of no-unused-vars-before-return
+  return group?.isCollapsed ? null : (
     <LayerContainer>
       {isRenameable && isLayerNamingEnabled ? (
-        <LayerInputWrapper>
-          <GroupIconsWrapper>
-            {GroupArrow}
-            <Icons.Group />
-          </GroupIconsWrapper>
+        <LayerInputWrapper isNested={isNested}>
+          <LayerIconWrapper>
+            <ShapeMaskWrapper element={element}>
+              <LayerIcon
+                element={element}
+                getProxiedUrl={getProxiedUrl}
+                currentPageBackgroundColor={currentPageBackgroundColor}
+              />
+            </ShapeMaskWrapper>
+          </LayerIconWrapper>
           <LayerInputForm onSubmit={handleSubmit}>
             <LayerInput
               tabIndex={-1}
@@ -476,63 +452,95 @@ function Group({ groupId }) {
         </LayerInputWrapper>
       ) : (
         <LayerButton
-          ref={groupRef}
-          id={groupDomId}
+          ref={layerRef}
+          id={layerId}
           onClick={handleClick}
           isSelected={isSelected}
+          isNested={isNested}
         >
-          <GroupIconsWrapper>
-            {GroupArrow}
-            <Icons.Group />
-          </GroupIconsWrapper>
+          <LayerIconWrapper>
+            <ShapeMaskWrapper element={element}>
+              <LayerIcon
+                element={element}
+                getProxiedUrl={getProxiedUrl}
+                currentPageBackgroundColor={currentPageBackgroundColor}
+              />
+            </ShapeMaskWrapper>
+          </LayerIconWrapper>
           <LayerDescription>
             <LayerContentContainer>
-              <LayerText>{group.name}</LayerText>
+              <LayerText>{layerName}</LayerText>
             </LayerContentContainer>
-            {group.isLocked && isLayerLockingEnabled && (
+            {element.isBackground && (
+              <IconWrapper>
+                <Icons.LockFilledClosed />
+              </IconWrapper>
+            )}
+            {element.isLocked && isLayerLockingEnabled && (
               <IconWrapper aria-label={__('Locked', 'web-stories')}>
-                <Icons.LockClosed />
+                {isNested ? <Icons.LockFilledClosed /> : <Icons.LockClosed />}
               </IconWrapper>
             )}
           </LayerDescription>
         </LayerButton>
       )}
-      {!isRenameable && (
+      {!element.isBackground && !isRenameable && (
         <ActionsContainer>
-          <Tooltip title={__('Delete Group', 'web-stories')} hasTail isDelayed>
+          {hasShapeMask && (
+            <Tooltip title={removeMaskTitle} hasTail isDelayed>
+              <LayerAction
+                aria-label={removeMaskTitle}
+                aria-describedby={layerId}
+                onClick={removeShapeMask}
+              >
+                <Icons.RemoveMask />
+              </LayerAction>
+            </Tooltip>
+          )}
+          <Tooltip title={__('Delete Layer', 'web-stories')} hasTail isDelayed>
             <LayerAction
               ref={deleteButtonRef}
               aria-label={__('Delete', 'web-stories')}
-              aria-describedby={groupDomId}
+              aria-describedby={layerId}
               onPointerDown={preventReorder}
-              onClick={() => deleteGroupAndElementsById({ groupId })}
+              onClick={() => deleteElementById({ elementId: element.id })}
             >
               <Icons.Trash />
             </LayerAction>
           </Tooltip>
-          <Tooltip
-            title={__('Duplicate Group', 'web-stories')}
-            hasTail
-            isDelayed
-          >
-            <LayerAction
-              aria-label={__('Duplicate', 'web-stories')}
-              aria-describedby={groupDomId}
-              onPointerDown={preventReorder}
-              onClick={handleDuplicateGroup}
+          {hasDuplicateMenu && (
+            <Tooltip
+              title={__('Duplicate Layer', 'web-stories')}
+              hasTail
+              isDelayed
             >
-              <Icons.PagePlus />
-            </LayerAction>
-          </Tooltip>
+              <LayerAction
+                aria-label={__('Duplicate', 'web-stories')}
+                aria-describedby={layerId}
+                onPointerDown={preventReorder}
+                onClick={() =>
+                  duplicateElementsById({ elementIds: [element.id] })
+                }
+              >
+                <Icons.PagePlus />
+              </LayerAction>
+            </Tooltip>
+          )}
           {isLayerLockingEnabled && (
             <Tooltip title={lockTitle} hasTail isDelayed>
               <LayerAction
                 aria-label={__('Lock/Unlock', 'web-stories')}
-                aria-describedby={groupDomId}
+                aria-describedby={layerId}
+                disabled={isNested}
                 onPointerDown={preventReorder}
-                onClick={handleLockGroup}
+                onClick={() =>
+                  updateElementById({
+                    elementId: element.id,
+                    properties: { isLocked: !element.isLocked },
+                  })
+                }
               >
-                <LockIcon />
+                {isNested ? <Icons.LockFilledClosed /> : <LockIcon />}
               </LayerAction>
             </Tooltip>
           )}
@@ -542,8 +550,8 @@ function Group({ groupId }) {
   );
 }
 
-Group.propTypes = {
-  groupId: PropTypes.string.isRequired,
+Layer.propTypes = {
+  element: StoryPropTypes.element.isRequired,
 };
 
-export default memo(Group);
+export default memo(Layer);
